requirements.txt
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
claude-api-py==0.0.3
click==8.1.6
exceptiongroup==1.1.2
fastapi==0.95.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
pydantic==1.10.12
python-dotenv==1.0.0
regex==2023.6.3
requests==2.31.0
sniffio==1.3.0
socksio==1.0.0
sseclient-py==1.7.2
starlette==0.27.0
tiktoken==0.4.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.22.0
